switch to output object
